requirements.txt
absl-py==1.3.0
alabaster==0.7.12
alembic==1.7.6
aniso8601==9.0.1
anyio==3.6.2
apipkg==3.0.1
appdirs
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.12.13
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
autocommand==2.2.2
awkward==1.10.2
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
Beaker==1.11.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
build==0.9.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.6
coverage==6.5.0
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
distro==1.8.0
dnspython==2.2.1
docutils==0.19
elasticsearch==7.9.0
email-validator==1.3.0
entrypoints==0.4
evdev==1.6.0
exceptiongroup==1.0.4
executing==1.2.0
extras==1.0.0
fastjsonschema==2.16.2
feedparser==6.0.10
filelock==3.8.2
fixtures==4.0.1
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SeaSurf==1.0.0
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==22.12.6
fonttools==4.38.0
future==0.18.2
gast==0.3.3
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
gssapi==1.8.2
h11==0.14.0
h5py==3.7.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.10.1
inflect==6.0.2
inflection==0.5.1
iniconfig==1.1.1
installer==0.6.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_packaging==0.12.3
jupyter_server==2.0.1
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
ldap3==2.9
lockfile==0.12.2
logutils==0.3.5
lxml==4.9.1
lz4==3.1.10
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
meson==0.64.1
mistune==2.0.4
mock==3.0.5
more-itertools==9.0.0
msgpack==1.0.4
mtcnn==0.1.1
multidict==6.0.3
Naked==0.1.31
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
nspektr==0.4.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
openshot-qt==3.0.0
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.11.1
parso==0.8.3
passlib==1.7.4
pbr==5.11.0
pdfarranger==1.9.2
pecan==1.4.2
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==6.2.5
Pillow==9.3.0
pkgconfig==1.5.5
platformdirs==2.6.0
pluggy==1.0.0
ply==3.11
portend==3.1.0
prettytable==3.5.0
progress==1.6
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==4.21.11
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyalpm==0.10.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.15.0
pycuda==2022.1
pydantic==1.10.2
pydot==1.4.2
pyee==8.2.2
Pygments==2.13.0
PyGObject==3.42.2
PyJWT==2.6.0
pylint==2.15.8
PyNaCl==1.4.0
PyOpenGL==3.1.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyppeteer==1.0.2
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyqtgraph==0.13.1
PyQtWebEngine==5.15.6
pyrsistent==0.19.2
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.4
pytools==2022.1.13
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
repoze.lru==0.7
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
resolvelib==0.9.0
retrying==1.3.3
rfc3987==1.3.8
rich==12.6.0
Routes==2.5.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
SCons==4.4.0
seaborn==0.12.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
shellescape==3.8.1
simplejson==3.18.0
six==1.16.0
smbus==1.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speaklater==1.3
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.24
sqlparse==0.4.3
sshtunnel==0.4.0
stack-data==0.6.2
tables==3.7.0
tabulate==0.9.0
TBB==0.2
team==1.0
tempora==5.1.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.7.0a0
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator
tesnorflow-gpu
termcolor==2.1.1
terminado==0.17.1
terminator==2.1.2
testpath==0.6.0
testtools==2.5.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.0
traitlets==5.7.0
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2022.12.1
typeguard==2.13.3
typing_extensions==4.4.0
uproot==4.3.5
uproot3-methods==0.10.1
uritemplate==4.1.1
urllib3==1.26.12
validate==5.1.0.dev0
validate-pyproject==0.10.1
virtualenv==20.17.1
waitress==2.1.2
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.4.2
websockets==10.3
WebTest==3.0.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
wxPython==4.2.0
xmlsec==1.3.13
zc.lockfile==2.0
zipp==3.11.0
zstandard==0.19.0
